+++ tomcat/src/main/java/org/apache/coyote/http/HttpContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.http;
 
 import java.util.stream.Stream;
