+++ tomcat/src/main/java/org/apache/coyote/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.http;
 
 import java.util.Arrays;
