+++ tomcat/src/main/java/org/apache/coyote/http/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.http;
 
 public enum HttpStatusCode {
