+++ tomcat/src/main/java/org/apache/coyote/http/request/HttpRequestHeader.java
@@ -1,8 +1,9 @@
-package org.apache.coyote;
+package org.apache.coyote.http.request;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.apache.coyote.http.HttpCookies;
 
 public class HttpRequestHeader {
