+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,9 +7,9 @@ import java.io.InputStreamReader;
 import java.net.Socket;
 import org.apache.catalina.handler.ServletRequestHandler;
 import org.apache.catalina.handler.ViewResolver;
-import org.apache.coyote.HttpRequest;
-import org.apache.coyote.HttpResponse;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
